Refactor tests and introduce Python model tests
