+++ examples/sushi/tests/test_order_items.yaml
@@ -1,0 +1,76 @@
+test_order_items:
+  model: sushi.order_items
+  inputs:
+    sushi.orders:
+      rows:
+        - id: 1
+          customer_id: 1
+          waiter_id: 1
+          start_ts: 2022-01-01 01:59:00
+          end_ts: 2022-01-01 02:29:00
+          ds: 2022-01-01
+        - id: 2
+          customer_id: 1
+          waiter_id: 2
+          start_ts: 2022-01-01 03:59:00
+          end_ts: 2022-01-01 03:29:00
+          ds: 2022-01-01
+    sushi.items:
+      rows:
+        - id: 1
+          name: ahi
+          price: 1.23
+          ds: 2022-01-01
+        - id: 2
+          name: aji
+          price: 2.34
+          ds: 2022-01-01
+        - id: 3
+          name: amaebi
+          price: 2.34
+          ds: 2022-01-01
+        - id: 4
+          name: anago
+          price: 2.34
+          ds: 2022-01-01
+        - id: 5
+          name: aoyagi
+          price: 2.34
+          ds: 2022-01-01
+  outputs:
+    query:
+      rows:
+      - id: 1
+        order_id: 1
+        item_id: 2
+        quantity: 4
+        ds: '2022-01-01'
+      - id: 2
+        order_id: 2
+        item_id: 3
+        quantity: 7
+        ds: '2022-01-01'
+      - id: 1
+        order_id: 2
+        item_id: 2
+        quantity: 8
+        ds: '2022-01-01'
+      - id: 0
+        order_id: 2
+        item_id: 1
+        quantity: 9
+        ds: '2022-01-01'
+      - id: 4
+        order_id: 2
+        item_id: 5
+        quantity: 9
+        ds: '2022-01-01'
+      - id: 3
+        order_id: 2
+        item_id: 4
+        quantity: 5
+        ds: '2022-01-01'
+  vars:
+    start: 2022-01-01
+    end: 2022-01-01
+    latest: 2022-01-01
